proc: call per file
