+++ c_src/sys/setproctitle.c
@@ -1,0 +1,42 @@
+/* Copyright (c) 2014, Michael Santos <michael.santos@gmail.com>
+ *
+ * Permission to use, copy, modify, and/or distribute this software for any
+ * purpose with or without fee is hereby granted, provided that the above
+ * copyright notice and this permission notice appear in all copies.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ * ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ */
+#include "alcove.h"
+#include "alcove_call.h"
+
+/*
+ * setproctitle(3)
+ */
+    ssize_t
+alcove_sys_setproctitle(alcove_state_t *ap, const char *arg, size_t len,
+        char *reply, size_t rlen)
+{
+#if defined(__linux__) || defined(__sunos__)
+    return alcove_mk_atom(reply, rlen, "undef");
+#else
+    int index = 0;
+
+    char name[MAXMSGLEN] = {0};
+    size_t nlen = sizeof(name);
+
+    if (alcove_decode_iolist(arg, len, &index, name, &nlen) < 0)
+        return -1;
+
+    (nlen == 0)
+        ? setproctitle(NULL)
+        : setproctitle("%s", name);
+
+    return alcove_mk_atom(reply, rlen, "ok");
+#endif
+}
